requirements.txt
flask==2.0.3
werkzeug==2.0.3
python-dotenv==1.0.0
openai==0.27.0
numpy==1.21.2
python-docx==0.8.11
scikit-learn==0.24.2